+++ CHANGES.rst
@@ -22,6 +22,7 @@ API relevant changes:
 * Reorderd the dimensions of comb_filters to time, freq, tau (#135)
 * `write_notes` uses `delimiter` instead of `sep` to seperate columns (#155)
 * `LSTMLayer` takes `Gate`s as arguments, all layers are callable (#161)
+* replaced `online` parameter of `FramedSignalProcessor` by `origin` (#169)
 
 Other changes: